refactor(text-image-morph): split into focused modules, trim comments
Break the two large files into small single-responsibility modules and cut the
verbose comments down to the non-obvious "why". No behaviour change.

- atlas.ts        glyph-atlas geometry from the paragraph
- layout.ts       word-wrapped page layout
- sampling.ts     image sampling pipeline (sample/prune/fit/assign)
- use-text-image-morph.ts  hook wiring layout + deferred sampling
- reveal.tsx      the Skia draw (atlas/transform/colour buffers) + letterPhase
- toggle-icon.tsx the FAB icon crossfade
- text-image-morph.tsx  the screen shell (Canvas + FAB)
