+++ src/animations/text-image-morph/index.ts
@@ -1,7 +1,5 @@
 export { TextImageMorph } from './text-image-morph';
 export { useTextImageMorph } from './use-text-image-morph';
-export type {
-  Atlas,
-  MorphTargets,
-  TextImageMorphData,
-} from './use-text-image-morph';
+export type { TextImageMorphData } from './use-text-image-morph';
+export type { Atlas } from './atlas';
+export type { MorphTargets } from './sampling';
